Color changes upon import, reorganized header code
